requirements.txt
anthropic>=0.40.0
tweepy>=4.14.0
python-telegram-bot>=20.7
beautifulsoup4>=4.12.0
requests>=2.31.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
tavily-python>=0.3.0
